1/11/17 Comitt
This includes most of the source code from Sauron’s theme as well as a
few changes I have made in the last day or so, including:
- removed branding header
- Using [caption] of featured image as the title under the grid thumbs
- Using $excerpt of post as on-hover text to ID tools/skills used in
project
- styled the on hover class SlideFromBottom by removing the icons.
- removed some elements from single-posts to show a less distracting
page without many of the meta info which is not crucial for the
projects page
- duplicated the home.php to projects.php as a template. This way I can
just use the Projects page as landing page. It also provides more
control as I can make changes to the forked php file.